# docs/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/daos/{slug}": {
      "get": {
        "operationId": "DaoController_detail",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaoDetailResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["daos"]
      }
    },
...
    "/v1/daos/{slug}/delegations": {
      "get": {
        "operationId": "DelegationsController_list",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Comma-delimited sort fields (prefix with - for desc)",
            "schema": {
              "example": "slug,-created_at",
              "type": "string"
            }
          },
          {
            "name": "dao",
            "required": false,
            "in": "query",
            "description": "Comma-delimited DAO slugs",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "description": "Comma-delimited proposal states",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source_type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "proposer",
            "required": false,
            "in": "query",
            "description": "0x-prefixed proposer address",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "binding",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "voting_starts_at_min",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "voting_starts_at_max",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DelegationListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["delegations"]
      }
    },
    "/v1/daos/{slug}/delegates/{delegate_address}/current": {
      "get": {
        "operationId": "DelegationsController_current",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "delegate_address",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "as_of_block_number",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentDelegatorsResponseDto"
                }
              }
            }
          },
          "301": {
            "description": "Redirect to canonical delegate address"
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["delegations"]
      }
    },
    "/v1/daos/{slug}/actors/{address}/delegation": {
      "get": {
        "operationId": "DelegationsController_actorDelegation",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActorDelegationResponseDto"
                }
              }
            }
          },
          "301": {
            "description": "Redirect to canonical actor address"
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["delegations"]
      }
    },
...
  },
  "info": {
    "title": "Kvorum API",
    "description": "Kvorum M1 DAO + proposal read API",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "type": "http",
        "description": "ADR-025 API-key bearer authentication"
      }
    },
    "schemas": {
...
      "ProblemDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "number"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "violations": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": ["type", "title", "status", "detail", "instance"]
      },
...
      "PaginationDto": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "number"
          },
          "next_cursor": {
            "type": "object",
            "nullable": true
          },
          "has_more": {
            "type": "boolean"
          }
        },
        "required": ["limit", "next_cursor", "has_more"]
      },
...
      "DaoLinksDto": {
        "type": "object",
        "properties": {
          "self": {
            "type": "string"
          }
        },
        "required": ["self"]
      },
      "DaoMetaDto": {
        "type": "object",
        "properties": {
          "last_updated_at": {
            "type": "string"
          },
          "links": {
            "$ref": "#/components/schemas/DaoLinksDto"
          }
        },
        "required": ["last_updated_at", "links"]
      },
...
      "DaoSourceDto": {
        "type": "object",
        "properties": {
          "source_type": {
            "type": "string"
          },
          "contract_address": {
            "type": "string"
          },
          "chain_id": {
            "type": "string"
          }
        },
        "required": ["source_type"]
      },
      "DaoDetailDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "website_url": {
            "type": "string"
          },
          "forum_url": {
            "type": "string"
          },
          "primary_token_address": {
            "type": "string"
          },
          "primary_chain_id": {
            "type": "string"
          },
          "_meta": {
            "$ref": "#/components/schemas/DaoMetaDto"
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DaoSourceDto"
            }
          }
        },
        "required": [
          "slug",
          "name",
          "description",
          "website_url",
          "forum_url",
          "primary_token_address",
          "primary_chain_id",
          "_meta",
          "sources"
        ]
      },
      "DaoDetailResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DaoDetailDto"
          }
        },
        "required": ["data"]
      },
...
      "DelegationActorLinksDto": {
        "type": "object",
        "properties": {
          "actor": {
            "type": "string"
          }
        },
        "required": ["actor"]
      },
      "DelegationActorMetaDto": {
        "type": "object",
        "properties": {
          "links": {
            "$ref": "#/components/schemas/DelegationActorLinksDto"
          }
        },
        "required": ["links"]
      },
      "DelegationActorDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          },
          "display_name": {
            "type": "object",
            "nullable": true
          },
          "_meta": {
            "$ref": "#/components/schemas/DelegationActorMetaDto"
          }
        },
        "required": ["address", "_meta"]
      },
      "DelegationListItemDto": {
        "type": "object",
        "properties": {
          "delegation_id": {
            "type": "string"
          },
          "delegator": {
            "$ref": "#/components/schemas/DelegationActorDto"
          },
          "delegate": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DelegationActorDto"
              }
            ]
          },
          "voting_power": {
            "type": "string"
          },
          "block_number": {
            "type": "string"
          },
          "event_type": {
            "type": "string"
          },
          "tx_hash": {
            "type": "string"
          },
          "created_at": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "delegation_id",
          "delegator",
          "delegate",
          "voting_power",
          "block_number",
          "event_type",
          "tx_hash",
          "created_at"
        ]
      },
      "DelegationListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DelegationListItemDto"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationDto"
          }
        },
        "required": ["data", "pagination"]
      },
      "CurrentDelegatorsMetaDto": {
        "type": "object",
        "properties": {
          "as_of_block_number": {
            "type": "string"
          }
        },
        "required": ["as_of_block_number"]
      },
      "CurrentDelegatorsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DelegationListItemDto"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationDto"
          },
          "_meta": {
            "$ref": "#/components/schemas/CurrentDelegatorsMetaDto"
          }
        },
        "required": ["data", "pagination", "_meta"]
      },
      "ActorDelegationResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DelegationListItemDto"
              }
            ]
          }
        },
        "required": ["data"]
      },
...
    }
  }
}